requirements.txt

httpx==0.23.0
pydantic-settings==2.1.0
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.4.2
Authlib==1.2.1
PyJWT==2.8.0
itsdangerous==2.1.2
pre_commit==3.7.0
